Door handle feature
Coded with Ari (adding co-author was glitching so mentionning it here)

Extracted more parameters from the model and added door handles to be returned. We sorted the data returned by the model such that door handles and doors are paired together in the model's dataframe response. We then output whether the door handle is to the left or to the right in the Approaching Phase with Tts.

Future improvements:
-Determine if we pull or push handle

Co-Authored-By: AriA700 <75389449+aria700@users.noreply.github.com>