MultiDoor Progress
We now detect and prompt the user to select from multiple doors. We introduced a new phase into the sequence.

It goes as follows:

Searching -> (if multiple doors are detected) Door Selection -> Calibrating -> Approaching

For a single door in the frame, the normal sequence of actions is used.

Searching -> Calibrating -> Approaching

Remaining things to do:
Have the user choose one of the doors and link that door to the Calibrating Phase.